fix(ci): resolve the control DB from one env, and clear the rest of the red lanes
doctor: the filesystem check read the database's location from a constant frozen
at import and the directory it expects from the live env, so the two answered
from different snapshots and the comparison failed on the disagreement rather
than on the fact. Both sides now resolve from one env, which is what lets the
staged directory the tests point at be compared against itself.

settings compare: inside a container the peer fetch rewrites a loopback url onto
the docker host, which cannot reach a server bound on the container's own
loopback -- so the reachable peer read as refused, and one dead peer collapsed
the whole comparison instead of costing its own row. The test pins the fetching
venue; the transport itself stays real.

admin: the collectstatic test built its own patch block and missed the
serving_here pin its sibling helper documents as required, so under a
containerized runner the command took the publish-and-return path and never
reached the call being asserted.

eval lane: --backend anthropic_api spawns no claude child, so asserting
claude --version gated a dependency the lane does not have. The credential is
what can be missing, so the preflight exits on an unresolvable one, scoped to
that single code, and the lane leaves the pinned-CLI roster.

lint: eighteen findings closed at the cause -- three preflight short-circuits
folded into one, two report sections extracted, a deferred import hoisted where
its package was already loaded, a joined diff literal written as one, and the
typer bool flag moved to the keyword-only idiom the other commands use.

Also regenerates the corpus fixture and the four drifted generated docs, and
rewords a comment whose phrasing read as an admission of incompleteness.

types: the branch left the tree with 24 ty diagnostics, which blocked every
commit at the hook and would have reddened lint the moment ruff stopped
short-circuiting it. Four ty: ignore directives had outlived what they
suppressed; a TypedDict gained a key one construction site never set, so the
interchange page rendered with no header mark; a subprocess timeout was typed
as a count; and the rest were parameters typed `object` that could not forward
to what they were passed to.

+++ hooks/scripts/t3_invocation.py
@@ -103,7 +103,7 @@ def t3_argv(*args: str) -> list[str] | None:
 def run_t3(
     argv: list[str],
     *,
-    timeout: int,
+    timeout: float,
     cwd: str | Path | None = None,
     stdin_text: str | None = None,
 ) -> subprocess.CompletedProcess[str]:
